SearchClient Wrapper 클래스 작성
파라미터가 너무 많아, 필요한 파라미터만 전달하여 기능하는 래퍼 클래스를 구현했다.
Symbolic Week와 Query, 혹은 TimeSpan과 Query를 전달하면 동작할 수 있다.

+++ batchwos-wsclient/src/main/java/com/siotman/batchwos/wsclient/search/domain/RetrieveParameters.java
@@ -8,8 +8,8 @@ import javax.xml.soap.SOAPException;
 
 @Data @Builder
 public class RetrieveParameters implements Parameters {
-    private Long firstRecord;
-    private Long count;
+    private Integer firstRecord;
+    private Integer count;
     private SortField sortField;
     private Option option;
